requirements.txt
apscheduler==3.6.1
dnspython==1.16.0
pymongo==3.8.0
azure-storage-blob==2.1.0
azure-cosmosdb-table==1.0.5
azure-common==1.1.21
azure-mgmt-storage==3.3.0
azure-storage-common==2.0.0
sendgrid==6.1.0
pika==1.0.1